refactor(bin-designer): split CompartmentEditor into focused sibling files
Decompose the 792-LOC CompartmentEditor.tsx so the orchestrator stays
under the 500-LOC ceiling. No behavior change.

- compartmentEditorColors.ts (61 LOC): pure color helpers — preview color
  sync (PREVIEW_COLOR_KEY, getPreviewColor) and HSL math
  (hexToHsl, getCompartmentFill, getCompartmentBorder). Wraps
  localStorage in try/catch to harden against privacy-mode failures.
- CompartmentEditorParts.tsx (256 LOC): GridCell + GhostPreview.
- CompartmentEditor.tsx (498 LOC): orchestrator + state.

Adds CompartmentEditorParts.test.tsx with 6 tests covering rendering,
dimension labels, selection state, and ghost preview action modes.
